Generating wheels only if running on Linux

+++ conda-dist/make_packages.sh
@@ -1,44 +1,48 @@
 #!/bin/bash
 
-CONDA_ENV='/home/giacomov/miniconda2'
+if [ "$(uname)" == "Linux" ]; then
 
-WHEEL2CONDA='/home/giacomov/.local/bin/wheel2conda'
-
-# activate conda
-source ${CONDA_ENV}/bin/activate
-
-echo "Using python:"
-which python
-
-echo "Make wheels"
-
-python make_wheels.py
-
-echo "Run wheel2conda"
-
-for wheel in *.whl
-do
+    CONDA_ENV='/home/giacomov/miniconda2'
     
-    echo Processing $wheel
+    WHEEL2CONDA='/home/giacomov/.local/bin/wheel2conda'
     
-    $WHEEL2CONDA $wheel
+    # activate conda
+    source ${CONDA_ENV}/bin/activate
     
-    # Remove all packages not for python 2.7
-    rm -rf linux-64/*py34*
-    rm -rf linux-64/*py35*
-    rm -rf osx-64/*py34*
-    rm -rf osx-64/*py35*
+    echo "Using python:"
+    which python
     
-done
+    echo "Make wheels"
+    
+    python make_wheels.py
+    
+    echo "Run wheel2conda"
+    
+    for wheel in *.whl
+    do
+        
+        echo Processing $wheel
+        
+        $WHEEL2CONDA $wheel
+        
+        # Remove all packages not for python 2.7
+        rm -rf linux-64/*py34*
+        rm -rf linux-64/*py35*
+        rm -rf osx-64/*py34*
+        rm -rf osx-64/*py35*
+        
+    done
+    
+    rm -rf win-32
+    rm -rf win-64
+    rm -rf linux-32
+    rm -rf *.whl
+    
+    echo "Add the wheels to the channel"
+    anaconda upload linux-64/* --force
+    anaconda upload osx-64/* --force
 
-rm -rf win-32
-rm -rf win-64
-rm -rf linux-32
-rm -rf *.whl
-
-echo "Add the wheels to the channel"
-anaconda upload linux-64/* --force
-anaconda upload osx-64/* --force
+fi
 
 echo "Make custom recipes"
 conda config --set anaconda_upload yes
